Modify find_rect to take an image, not a file_path
This is to move away from saving templates. With the ability to find and locate the squares that would be used to track motion now, we don't need to save templates. And even if we do, we can always read the iamge first before passing it into the function. This allows for more flexibility.