Account for imcrop
When using imcrop, the size of the image given is always 1 more in both dimension
This meant that after calculating some height/width, we would always have to add 1
after an imcrop. Now we just have imcrop crop an image with a height/width 1 less
than what we previously calculated. This way the resulting image is the expected size.