Readd phase correlation, fix logitcal error in normxcorr
Previously, when I performed template matching in an interpolated image,
I returned the coordinates as relative to the original image; however, I did not do the same,
for an uninterpolated image. This fixes that.

Also, even though phase correlation failed last time, I now understand
the code more and as such can see that the reason why it failed may be
because I converted the coordinate incorrectly. I'm trying to see if that really
was the reason.